add authentication with github and jwt

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from 'zod';
 import { prisma } from '../lib/prisma';
 
 export async function memoryRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,7 +26,7 @@ export async function memoryRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -29,6 +36,10 @@ export async function memoryRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -47,14 +58,14 @@ export async function memoryRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '3162bcec-fb95-47b8-bbaf-0aa872d6ca0d',
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -68,7 +79,17 @@ export async function memoryRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params);
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -80,12 +101,22 @@ export async function memoryRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: { id },
